DO NOT MAKE CHANGES IN THIS FILE
This is the older version, the version I had uploaded before the call on 28th. If you guys wanna make any changes, do it to the HotelSoftwarev2.c (the one without brackets around v2) as that is the version after the call. Really sorry if the file names are getting confusing. I'll try and delete the files not in usage.